:truck: generalize the useFocusTrap hook and move it to a generic location
