+++ src/utils/useFocusTrap/index.js
@@ -1,0 +1,2 @@
+import useFocusTrap from './useFocusTrap';
+export default useFocusTrap;
